fix the disable aspects in the bottons of reschedule and cancel
